scroll-animations/css/at-scroll-timeline-before-phase.html

<!DOCTYPE html>
<link rel="help" src="https://drafts.csswg.org/scroll-animations-1/#scroll-timeline-at-rule">
<link rel="help" src="https://drafts.csswg.org/scroll-animations-1/#phase-algorithm">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/web-animations/testcommon.js"></script>
<style>
  #scroller {
    overflow: scroll;
    width: 100px;
    height: 100px;
  }
  #contents {
    height: 200px;
  }
  @keyframes expand {
    from { width: 100px; }
    to { width: 200px; }
  }
  @scroll-timeline timeline {
    source: selector(#scroller);
    time-range: 10s;
    start: 50px;
    end: 100px;
  }
  #element {
    width: 0px;
    animation: expand 10s linear;
    animation-timeline: timeline;
  }
  /* Ensure stable expectations if feature is not supported */
  @supports not (animation-timeline:foo) {
    #element { animation-play-state: paused; }
  }
</style>
<div id=scroller>
  <div id=contents></div>
</div>
<div id=element></div>
<script>
  promise_test(async (t) => {
    // Animation should not apply yet.
    assert_equals(getComputedStyle(element).width, '0px');
    await waitForFrame();
    // Animation should still not apply.
    assert_equals(getComputedStyle(element).width, '0px');
  }, 'Animation does not apply in before phase');
</script>
